backend/src/main/java/group25/group25/task/serviceimplementation/TaskServiceImpl.java
package group25.group25.task.serviceimplementation;


import group25.group25.task.model.Task;
import group25.group25.task.repository.TaskRepository;
import group25.group25.task.service.TaskService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;

@Component
public class TaskServiceImpl implements TaskService {

    @Autowired
    TaskRepository taskRepository;



    @Override
    public Task saveTask(Task task) {

        return  taskRepository.save(task);

    }


}
